requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
astroid==2.1.0            # via pylint
autopep8==1.4
boto3==1.9.75             # via smart-open
boto==2.49.0              # via smart-open
botocore==1.12.75         # via boto3, s3transfer
bz2file==0.98             # via smart-open
certifi==2018.11.29       # via requests
chardet==3.0.4            # via requests
click==7.0                # via flask, pip-tools
colorama==0.4.1           # via pylint
cycler==0.10.0            # via matplotlib
docutils==0.14            # via botocore
flask==1.0.2
gensim==3.6.0
idna==2.8                 # via requests
isort==4.3.4              # via pylint
itsdangerous==1.1.0       # via flask
jinja2==2.10              # via flask
jmespath==0.9.3           # via boto3, botocore
kiwisolver==1.0.1         # via matplotlib
lazy-object-proxy==1.3.1  # via astroid
markupsafe==1.1.0         # via jinja2
matplotlib==3.0.2
mccabe==0.6.1             # via pylint
nltk==3.4                 # via textblob
numpy==1.15.4             # via gensim, matplotlib, pandas, scikit-learn, scipy, spectral
pandas==0.23.4
pip-tools==3.2.0
pipdeptree==0.13.1
pycodestyle==2.4.0        # via autopep8
pylint==2.1.1
pyparsing==2.3.0          # via matplotlib
python-dateutil==2.7.5    # via botocore, matplotlib, pandas
pytz==2018.9              # via pandas
requests==2.21.0          # via smart-open
s3transfer==0.1.13        # via boto3
scikit-learn==0.20.0
scipy==1.2.0              # via gensim, scikit-learn
singledispatch==3.4.0.3   # via nltk
six==1.12.0               # via astroid, cycler, gensim, nltk, pip-tools, python-dateutil, singledispatch
smart-open==1.7.1         # via gensim
spectral==0.19
textblob==0.15.1
urllib3==1.24.1           # via botocore, requests
werkzeug==0.14.1          # via flask
wrapt==1.10.11            # via astroid
